Allowing non-primitive with fixed types
This change now allows non-primitive encoding to work with fixed types
on either side of the bridge.

Note that if the ros type specified mqtt2ros does not match the one
received from the mqtt type topic then the config file specified version
will win (eg, the type will not be allowed to change).

Note that this is necessary since ROS doesn't really allow us to change
the type of topics on the fly and since we used a fixed type, the
publisher was created on startup.